Break up single file mod
